Default combobox values bug fixed.
There was a bug which caused the default values in the
edit form's id comboboxes to be invalid. It had three
main reasons: the fields values weren't properly concatened,
some forms has the wrong definition and the constant values
were loaded after the first value of the combobox, so they
could set some fields to null if they hadn't been selected
by the user.