Record deleting button added.
A new button has been added, which allows us to
delete the currently selected record.
Also made some minor changes on the DAO, and
added the new strings translations.
Please bear in mind this is a first version,
and it will probably crash in some forms
(I have to test it further).